requirements.txt
sqlalchemy==2.0.31
pymysql==1.1.0
azure-identity==1.16.0
python-dotenv==1.0.1
azure-cli